pyproject.toml
[project]
name = "bwki2025-lowres-langid-backend"
version = "0.1.0"
description = "Language identification backend using FastText"
requires-python = ">=3.11"
dependencies = [
    "langcodes",
    "fasttext-numpy2",
    "pyicu",
    "typing-extensions",
    "pydantic-settings",
    "starlette",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "numpy>=1.21.0",
    "requests>=2.28.0",
    "filelock>=3.19.1",
    "huggingface-hub>=0.35.0",
    "hf-transfer>=0.1.9",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = []
